# Optimize bitcount command by SIMD