--- a/src/bitops.c
+++ b/src/bitops.c
@@ -28,27 +28,112 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
+#define HAVE_AVX2
 #include "server.h"
-
+#ifdef HAVE_AVX2
+#include <immintrin.h>
+#endif
 /* -----------------------------------------------------------------------------
  * Helpers and low level bit functions.
  * -------------------------------------------------------------------------- */
+
+static const unsigned char bitsinbyte[256] = {
+    0, 1, 1, 2, 1, 2, 2, 3, 1, 2, 2, 3, 2, 3, 3, 4, 1, 2, 2, 3, 2, 3, 3, 4, 2, 3, 3, 4, 3, 4, 4, 5, 1, 2, 2, 3, 2,
+    3, 3, 4, 2, 3, 3, 4, 3, 4, 4, 5, 2, 3, 3, 4, 3, 4, 4, 5, 3, 4, 4, 5, 4, 5, 5, 6, 1, 2, 2, 3, 2, 3, 3, 4, 2, 3,
+    3, 4, 3, 4, 4, 5, 2, 3, 3, 4, 3, 4, 4, 5, 3, 4, 4, 5, 4, 5, 5, 6, 2, 3, 3, 4, 3, 4, 4, 5, 3, 4, 4, 5, 4, 5, 5,
+    6, 3, 4, 4, 5, 4, 5, 5, 6, 4, 5, 5, 6, 5, 6, 6, 7, 1, 2, 2, 3, 2, 3, 3, 4, 2, 3, 3, 4, 3, 4, 4, 5, 2, 3, 3, 4,
+    3, 4, 4, 5, 3, 4, 4, 5, 4, 5, 5, 6, 2, 3, 3, 4, 3, 4, 4, 5, 3, 4, 4, 5, 4, 5, 5, 6, 3, 4, 4, 5, 4, 5, 5, 6, 4,
+    5, 5, 6, 5, 6, 6, 7, 2, 3, 3, 4, 3, 4, 4, 5, 3, 4, 4, 5, 4, 5, 5, 6, 3, 4, 4, 5, 4, 5, 5, 6, 4, 5, 5, 6, 5, 6,
+    6, 7, 3, 4, 4, 5, 4, 5, 5, 6, 4, 5, 5, 6, 5, 6, 6, 7, 4, 5, 5, 6, 5, 6, 6, 7, 5, 6, 6, 7, 6, 7, 7, 8};
+
+#ifdef HAVE_AVX2
+/* The SIMD version of popcount enhances performance through parallel lookup tables. */
+ATTRIBUTE_TARGET_AVX2
+static long long serverPopcountAVX2(void *s, long count) {
+    long i = 0;
+    unsigned char *p = (unsigned char *)s;
+    long long bits = 0;
+
+    const __m256i lookup = _mm256_setr_epi8(
+        /* 0 */ 0, /* 1 */ 1, /* 2 */ 1, /* 3 */ 2,
+        /* 4 */ 1, /* 5 */ 2, /* 6 */ 2, /* 7 */ 3,
+        /* 8 */ 1, /* 9 */ 2, /* a */ 2, /* b */ 3,
+        /* c */ 2, /* d */ 3, /* e */ 3, /* f */ 4,
+        /* 0 */ 0, /* 1 */ 1, /* 2 */ 1, /* 3 */ 2,
+        /* 4 */ 1, /* 5 */ 2, /* 6 */ 2, /* 7 */ 3,
+        /* 8 */ 1, /* 9 */ 2, /* a */ 2, /* b */ 3,
+        /* c */ 2, /* d */ 3, /* e */ 3, /* f */ 4);
+
+    const __m256i low_mask = _mm256_set1_epi8(0x0f);
+    __m256i acc = _mm256_setzero_si256();
+
+/* Count 32 bytes per iteration*/
+#define ITER_32_BYTES                                                             \
+    {                                                                             \
+        const __m256i vec = _mm256_loadu_si256((const __m256i *)(p + i));         \
+        const __m256i lo = _mm256_and_si256(vec, low_mask);                       \
+        const __m256i hi = _mm256_and_si256(_mm256_srli_epi16(vec, 4), low_mask); \
+        const __m256i popcnt1 = _mm256_shuffle_epi8(lookup, lo);                  \
+        const __m256i popcnt2 = _mm256_shuffle_epi8(lookup, hi);                  \
+        local = _mm256_add_epi8(local, popcnt1);                                  \
+        local = _mm256_add_epi8(local, popcnt2);                                  \
+        i += 32;                                                                  \
+    }
+
+    /* We divide the array into the following three parts
+    *        Part A         Part B       Part C
+    * +-----------------+--------------+---------+
+    * | 8 * 32bytes * X |  32bytes * Y | Z bytes |
+    * +-----------------+--------------+---------+
+    * /
+
+    /* Parrt A: loop unrolling, processing 8 * 32 bytes per iteration */
+    while (i + 8 * 32 <= count) {
+        __m256i local = _mm256_setzero_si256();
+        ITER_32_BYTES ITER_32_BYTES ITER_32_BYTES ITER_32_BYTES
+        ITER_32_BYTES ITER_32_BYTES ITER_32_BYTES ITER_32_BYTES
+        acc = _mm256_add_epi64(acc, _mm256_sad_epu8(local, _mm256_setzero_si256()));
+    }
+
+    /* Part B: when the remaining data length is less than 8 * 32 bytes,
+     * process 32 bytes per iteration. */    
+    __m256i local = _mm256_setzero_si256();
+    while (i + 32 <= count) {
+        ITER_32_BYTES;
+    }
+    acc = _mm256_add_epi64(acc, _mm256_sad_epu8(local, _mm256_setzero_si256()));
+
+#undef ITER_32_BYTES
+
+    bits += _mm256_extract_epi64(acc, 0);
+    bits += _mm256_extract_epi64(acc, 1);
+    bits += _mm256_extract_epi64(acc, 2);
+    bits += _mm256_extract_epi64(acc, 3);
+
+    /* Part C: count the remaining bytes. */
+    for (; i < count; i++) {
+        bits += bitsinbyte[p[i]];
+    }
+
+    return bits;
+}
+#endif
 
 /* Count number of bits set in the binary array pointed by 's' and long
  * 'count' bytes. The implementation of this function is required to
  * work with an input string length up to 512 MB or more (server.proto_max_bulk_len) */
 long long serverPopcount(void *s, long count) {
+#ifdef HAVE_AVX2
+    /* If length of s >= 256 bits and the CPU supports AVX2,
+     * we prefer to use the SIMD version */
+    if (count >= 32) {
+        return serverPopcountAVX2(s, count);
+    }
+#endif
+
     long long bits = 0;
     unsigned char *p = s;
     uint32_t *p4;
-    static const unsigned char bitsinbyte[256] = {
-        0, 1, 1, 2, 1, 2, 2, 3, 1, 2, 2, 3, 2, 3, 3, 4, 1, 2, 2, 3, 2, 3, 3, 4, 2, 3, 3, 4, 3, 4, 4, 5, 1, 2, 2, 3, 2,
-        3, 3, 4, 2, 3, 3, 4, 3, 4, 4, 5, 2, 3, 3, 4, 3, 4, 4, 5, 3, 4, 4, 5, 4, 5, 5, 6, 1, 2, 2, 3, 2, 3, 3, 4, 2, 3,
-        3, 4, 3, 4, 4, 5, 2, 3, 3, 4, 3, 4, 4, 5, 3, 4, 4, 5, 4, 5, 5, 6, 2, 3, 3, 4, 3, 4, 4, 5, 3, 4, 4, 5, 4, 5, 5,
-        6, 3, 4, 4, 5, 4, 5, 5, 6, 4, 5, 5, 6, 5, 6, 6, 7, 1, 2, 2, 3, 2, 3, 3, 4, 2, 3, 3, 4, 3, 4, 4, 5, 2, 3, 3, 4,
-        3, 4, 4, 5, 3, 4, 4, 5, 4, 5, 5, 6, 2, 3, 3, 4, 3, 4, 4, 5, 3, 4, 4, 5, 4, 5, 5, 6, 3, 4, 4, 5, 4, 5, 5, 6, 4,
-        5, 5, 6, 5, 6, 6, 7, 2, 3, 3, 4, 3, 4, 4, 5, 3, 4, 4, 5, 4, 5, 5, 6, 3, 4, 4, 5, 4, 5, 5, 6, 4, 5, 5, 6, 5, 6,
-        6, 7, 3, 4, 4, 5, 4, 5, 5, 6, 4, 5, 5, 6, 5, 6, 6, 7, 4, 5, 5, 6, 5, 6, 6, 7, 5, 6, 6, 7, 6, 7, 7, 8};
 
     /* Count initial bytes not aligned to 32 bit. */
     while ((unsigned long)p & 3 && count) {