--- a/src/unit/test_bitops.c
+++ b/src/unit/test_bitops.c
@@ -1,4 +1,5 @@
 #include <time.h>
+#include <stdint.h>
 
 #include "test_help.h"