@xbasel 2025-02-17
Can I add tests to make sure the same results are achieved, with, and without AVX?