@chzhoo 2025-02-17
> > @xbasel Sure! If there is anything I can help, please let me know.
> 
> there was a typo :), I asked whether you can add tests (unless you think the existing tests are enough)

No problem, I will review the tests to ensure they are enough and get back to you.