@chzhoo 2025-02-18
> Can you add tests to make sure the same results are achieved, with, and without AVX?

Test code to check scalar/vector result consistency has been committed.